--- gene-pool-backend/Transcriber.cs
@@ -1,838 +1,0 @@
-﻿//
-// Copyright (c) Microsoft. All rights reserved.
-// Licensed under the MIT license. See LICENSE.md file in the project root for full license information.
-//
-
-// <toplevel>
-using System;
-using System.IO;
-using System.Threading.Tasks;
-using Microsoft.CognitiveServices.Speech;
-using Microsoft.CognitiveServices.Speech.Audio;
-// </toplevel>
-
-namespace gene_pool_backend {
-  public class Transcribers {
-    // Speech recognition from microphone.
-    public static async Task RecognitionWithMicrophoneAsync() {
-      // <recognitionWithMicrophone>
-      // Creates an instance of a speech config with specified subscription key and service region.
-      // Replace with your own subscription key and service region (e.g., "westus").
-      // The default language is "en-us".
-      var config = SpeechConfig.FromSubscription("YourSubscriptionKey", "YourServiceRegion");
-
-      // Creates a speech recognizer using microphone as audio input.
-      using (var recognizer = new SpeechRecognizer(config)) {
-        // Starts recognizing.
-        Console.WriteLine("Say something...");
-
-        // Starts speech recognition, and returns after a single utterance is recognized. The end of a
-        // single utterance is determined by listening for silence at the end or until a maximum of 15
-        // seconds of audio is processed.  The task returns the recognition text as result.
-        // Note: Since RecognizeOnceAsync() returns only a single utterance, it is suitable only for single
-        // shot recognition like command or query.
-        // For long-running multi-utterance recognition, use StartContinuousRecognitionAsync() instead.
-        var result = await recognizer.RecognizeOnceAsync().ConfigureAwait(false);
-
-        // Checks result.
-        if (result.Reason == ResultReason.RecognizedSpeech) {
-          Console.WriteLine($"RECOGNIZED: Text={result.Text}");
-        } else if (result.Reason == ResultReason.NoMatch) {
-          Console.WriteLine($"NOMATCH: Speech could not be recognized.");
-        } else if (result.Reason == ResultReason.Canceled) {
-          var cancellation = CancellationDetails.FromResult(result);
-          Console.WriteLine($"CANCELED: Reason={cancellation.Reason}");
-
-          if (cancellation.Reason == CancellationReason.Error) {
-            Console.WriteLine($"CANCELED: ErrorCode={cancellation.ErrorCode}");
-            Console.WriteLine($"CANCELED: ErrorDetails={cancellation.ErrorDetails}");
-            Console.WriteLine($"CANCELED: Did you update the subscription info?");
-          }
-        }
-      }
-      // </recognitionWithMicrophone>
-    }
-
-    // Speech recognition in the specified spoken language and uses detailed output format.
-    public static async Task RecognitionWithLanguageAndDetailedOutputAsync() {
-      // Creates an instance of a speech config with specified subscription key and service region.
-      // Replace with your own subscription key and service region (e.g., "westus").
-      var config = SpeechConfig.FromSubscription("YourSubscriptionKey", "YourServiceRegion");
-
-      // Replace the language with your language in BCP-47 format, e.g., en-US.
-      var language = "de-DE";
-      config.OutputFormat = OutputFormat.Detailed;
-
-      // Creates a speech recognizer for the specified language, using microphone as audio input.
-      // Requests detailed output format.
-      using (var recognizer = new SpeechRecognizer(config, language)) {
-        // Starts recognizing.
-        Console.WriteLine($"Say something in {language} ...");
-
-        // Starts speech recognition, and returns after a single utterance is recognized. The end of a
-        // single utterance is determined by listening for silence at the end or until a maximum of 15
-        // seconds of audio is processed.  The task returns the recognition text as result.
-        // Note: Since RecognizeOnceAsync() returns only a single utterance, it is suitable only for single
-        // shot recognition like command or query.
-        // For long-running multi-utterance recognition, use StartContinuousRecognitionAsync() instead.
-        var result = await recognizer.RecognizeOnceAsync().ConfigureAwait(false);
-
-        // Checks result.
-        if (result.Reason == ResultReason.RecognizedSpeech) {
-          Console.WriteLine($"RECOGNIZED: Text={result.Text}");
-          Console.WriteLine("  DETAILED RESULTS:");
-
-          var detailedResults = result.Best();
-          foreach (var item in detailedResults) // NOTE: We need to put this in all languages, or take it out of CSharp
-          {
-            Console.WriteLine($"    Confidence: {item.Confidence}, Text: {item.Text}, LexicalForm: {item.LexicalForm}, NormalizedForm: {item.NormalizedForm}, MaskedNormalizedForm: {item.MaskedNormalizedForm}");
-          }
-        } else if (result.Reason == ResultReason.NoMatch) {
-          Console.WriteLine($"NOMATCH: Speech could not be recognized.");
-        } else if (result.Reason == ResultReason.Canceled) {
-          var cancellation = CancellationDetails.FromResult(result);
-          Console.WriteLine($"CANCELED: Reason={cancellation.Reason}");
-
-          if (cancellation.Reason == CancellationReason.Error) {
-            Console.WriteLine($"CANCELED: ErrorCode={cancellation.ErrorCode}");
-            Console.WriteLine($"CANCELED: ErrorDetails={cancellation.ErrorDetails}");
-            Console.WriteLine($"CANCELED: Did you update the subscription info?");
-          }
-        }
-      }
-    }
-
-    // Speech recognition using a customized model.
-    public static async Task RecognitionUsingCustomizedModelAsync() {
-      // <recognitionCustomized>
-      // Creates an instance of a speech config with specified subscription key and service region.
-      // Replace with your own subscription key and service region (e.g., "westus").
-      var config = SpeechConfig.FromSubscription("YourSubscriptionKey", "YourServiceRegion");
-      // Replace with the CRIS endpoint id of your customized model.
-      var sourceLanguageConfig = SourceLanguageConfig.FromLanguage("en-US", "YourEndpointId");
-
-      // Creates a speech recognizer using microphone as audio input.
-      using (var recognizer = new SpeechRecognizer(config, sourceLanguageConfig)) {
-        Console.WriteLine("Say something...");
-
-        // Starts speech recognition, and returns after a single utterance is recognized. The end of a
-        // single utterance is determined by listening for silence at the end or until a maximum of 15
-        // seconds of audio is processed.  The task returns the recognition text as result.
-        // Note: Since RecognizeOnceAsync() returns only a single utterance, it is suitable only for single
-        // shot recognition like command or query.
-        // For long-running multi-utterance recognition, use StartContinuousRecognitionAsync() instead.
-        var result = await recognizer.RecognizeOnceAsync().ConfigureAwait(false);
-
-        // Checks results.
-        if (result.Reason == ResultReason.RecognizedSpeech) {
-          Console.WriteLine($"RECOGNIZED: Text={result.Text}");
-        } else if (result.Reason == ResultReason.NoMatch) {
-          Console.WriteLine($"NOMATCH: Speech could not be recognized.");
-        } else if (result.Reason == ResultReason.Canceled) {
-          var cancellation = CancellationDetails.FromResult(result);
-          Console.WriteLine($"CANCELED: Reason={cancellation.Reason}");
-
-          if (cancellation.Reason == CancellationReason.Error) {
-            Console.WriteLine($"CANCELED: ErrorCode={cancellation.ErrorCode}");
-            Console.WriteLine($"CANCELED: ErrorDetails={cancellation.ErrorDetails}");
-            Console.WriteLine($"CANCELED: Did you update the subscription info?");
-          }
-        }
-      }
-      // </recognitionCustomized>
-    }
-
-    // Continuous speech recognition.
-    public static async Task ContinuousRecognitionWithFileAsync() {
-      // <recognitionContinuousWithFile>
-      // Creates an instance of a speech config with specified subscription key and service region.
-      // Replace with your own subscription key and service region (e.g., "westus").
-      var config = SpeechConfig.FromSubscription("YourSubscriptionKey", "YourServiceRegion");
-
-      var stopRecognition = new TaskCompletionSource<int>();
-
-      // Creates a speech recognizer using file as audio input.
-      // Replace with your own audio file name.
-      using (var audioInput = AudioConfig.FromWavFileInput("test2.wav")) {
-        using (var recognizer = new SpeechRecognizer(config, audioInput)) {
-          // Subscribes to events.
-          recognizer.Recognizing += (s, e) =>
-          {
-            Console.WriteLine($"RECOGNIZING: Text={e.Result.Text}");
-          };
-
-          recognizer.Recognized += (s, e) =>
-          {
-            if (e.Result.Reason == ResultReason.RecognizedSpeech) {
-              Console.WriteLine($"RECOGNIZED: Text={e.Result.Text}");
-            } else if (e.Result.Reason == ResultReason.NoMatch) {
-              Console.WriteLine($"NOMATCH: Speech could not be recognized.");
-            }
-          };
-
-          recognizer.Canceled += (s, e) =>
-          {
-            Console.WriteLine($"CANCELED: Reason={e.Reason}");
-
-            if (e.Reason == CancellationReason.Error) {
-              Console.WriteLine($"CANCELED: ErrorCode={e.ErrorCode}");
-              Console.WriteLine($"CANCELED: ErrorDetails={e.ErrorDetails}");
-              Console.WriteLine($"CANCELED: Did you update the subscription info?");
-            }
-
-            stopRecognition.TrySetResult(0);
-          };
-
-          recognizer.SessionStarted += (s, e) =>
-          {
-            Console.WriteLine("\n    Session started event.");
-          };
-
-          recognizer.SessionStopped += (s, e) =>
-          {
-            Console.WriteLine("\n    Session stopped event.");
-            Console.WriteLine("\nStop recognition.");
-            stopRecognition.TrySetResult(0);
-          };
-
-          // Starts continuous recognition. Uses StopContinuousRecognitionAsync() to stop recognition.
-          await recognizer.StartContinuousRecognitionAsync().ConfigureAwait(false);
-
-          // Waits for completion.
-          // Use Task.WaitAny to keep the task rooted.
-          Task.WaitAny(new[] { stopRecognition.Task });
-
-          // Stops recognition.
-          await recognizer.StopContinuousRecognitionAsync().ConfigureAwait(false);
-        }
-      }
-      // </recognitionContinuousWithFile>
-    }
-
-    public static async Task SpeechRecognitionWithCompressedInputPullStreamAudio() {
-      // <recognitionWithCompressedInputPullStreamAudio>
-      // Creates an instance of a speech config with specified subscription key and service region.
-      // Replace with your own subscription key and service region (e.g., "westus").
-      var config = SpeechConfig.FromSubscription("YourSubscriptionKey", "YourServiceRegion");
-
-      var stopRecognition = new TaskCompletionSource<int>();
-
-      // Create an audio stream from a wav file.
-      // Replace with your own audio file name.
-
-      using (var audioInput = AudioConfig.FromStreamInput(new PullAudioInputStream(new BinaryAudioStreamReader(
-                              new BinaryReader(File.OpenRead(Path.GetFullPath("test2.mp3")))),
-                              AudioStreamFormat.GetCompressedFormat(AudioStreamContainerFormat.MP3)))) {
-        // Creates a speech recognizer using audio stream input.
-        using (var recognizer = new SpeechRecognizer(config, audioInput)) {
-          // Subscribes to events.
-          recognizer.Recognizing += (s, e) =>
-          {
-            Console.WriteLine($"RECOGNIZING: Text={e.Result.Text}");
-          };
-
-          recognizer.Recognized += (s, e) =>
-          {
-            if (e.Result.Reason == ResultReason.RecognizedSpeech) {
-              Console.WriteLine($"RECOGNIZED: Text={e.Result.Text}");
-            } else if (e.Result.Reason == ResultReason.NoMatch) {
-              Console.WriteLine($"NOMATCH: Speech could not be recognized.");
-            }
-          };
-
-          recognizer.Canceled += (s, e) =>
-          {
-            Console.WriteLine($"CANCELED: Reason={e.Reason}");
-
-            if (e.Reason == CancellationReason.Error) {
-              Console.WriteLine($"CANCELED: ErrorCode={e.ErrorCode}");
-              Console.WriteLine($"CANCELED: ErrorDetails={e.ErrorDetails}");
-              Console.WriteLine($"CANCELED: Did you update the subscription info?");
-            }
-
-            stopRecognition.TrySetResult(0);
-          };
-
-          recognizer.SessionStarted += (s, e) =>
-          {
-            Console.WriteLine("\nSession started event.");
-          };
-
-          recognizer.SessionStopped += (s, e) =>
-          {
-            Console.WriteLine("\nSession stopped event.");
-            Console.WriteLine("\nStop recognition.");
-            stopRecognition.TrySetResult(0);
-          };
-
-          // Starts continuous recognition. Uses StopContinuousRecognitionAsync() to stop recognition.
-          await recognizer.StartContinuousRecognitionAsync().ConfigureAwait(false);
-
-          // Waits for completion.
-          // Use Task.WaitAny to keep the task rooted.
-          Task.WaitAny(new[] { stopRecognition.Task });
-
-          // Stops recognition.
-          await recognizer.StopContinuousRecognitionAsync().ConfigureAwait(false);
-        }
-      }
-      // </recognitionWithCompressedInputPullStreamAudio>
-    }
-
-    public static async Task SpeechRecognitionWithCompressedInputPushStreamAudio() {
-      // <recognitionWithCompressedInputPushStreamAudio>
-      // Creates an instance of a speech config with specified subscription key and service region.
-      // Replace with your own subscription key and service region (e.g., "westus").
-      var config = SpeechConfig.FromSubscription("YourSubscriptionKey", "YourServiceRegion");
-
-      var stopRecognition = new TaskCompletionSource<int>();
-
-      using (var pushStream = AudioInputStream.CreatePushStream(AudioStreamFormat.GetCompressedFormat(AudioStreamContainerFormat.MP3))) {
-        using (var audioInput = AudioConfig.FromStreamInput(pushStream)) {
-          // Creates a speech recognizer using audio stream input.
-          using (var recognizer = new SpeechRecognizer(config, audioInput)) {
-            // Subscribes to events.
-            recognizer.Recognizing += (s, e) =>
-            {
-              Console.WriteLine($"RECOGNIZING: Text={e.Result.Text}");
-            };
-
-            recognizer.Recognized += (s, e) =>
-            {
-              if (e.Result.Reason == ResultReason.RecognizedSpeech) {
-                Console.WriteLine($"RECOGNIZED: Text={e.Result.Text}");
-              } else if (e.Result.Reason == ResultReason.NoMatch) {
-                Console.WriteLine($"NOMATCH: Speech could not be recognized.");
-              }
-            };
-
-            recognizer.Canceled += (s, e) =>
-            {
-              Console.WriteLine($"CANCELED: Reason={e.Reason}");
-
-              if (e.Reason == CancellationReason.Error) {
-                Console.WriteLine($"CANCELED: ErrorCode={e.ErrorCode}");
-                Console.WriteLine($"CANCELED: ErrorDetails={e.ErrorDetails}");
-                Console.WriteLine($"CANCELED: Did you update the subscription info?");
-              }
-
-              stopRecognition.TrySetResult(0);
-            };
-
-            recognizer.SessionStarted += (s, e) =>
-            {
-              Console.WriteLine("\nSession started event.");
-            };
-
-            recognizer.SessionStopped += (s, e) =>
-            {
-              Console.WriteLine("\nSession stopped event.");
-              Console.WriteLine("\nStop recognition.");
-              stopRecognition.TrySetResult(0);
-            };
-
-            // Starts continuous recognition. Uses StopContinuousRecognitionAsync() to stop recognition.
-            await recognizer.StartContinuousRecognitionAsync().ConfigureAwait(false);
-
-            using (BinaryAudioStreamReader reader = Utility.CreateBinaryFileReader(Path.GetFullPath("test2.mp3"))) {
-              byte[] buffer = new byte[1000];
-              while (true) {
-                var readSamples = reader.Read(buffer, (uint)buffer.Length);
-                if (readSamples == 0) {
-                  break;
-                }
-                pushStream.Write(buffer, readSamples);
-              }
-            }
-            pushStream.Close();
-
-            // Waits for completion.
-            // Use Task.WaitAny to keep the task rooted.
-            Task.WaitAny(new[] { stopRecognition.Task });
-
-            // Stops recognition.
-            await recognizer.StopContinuousRecognitionAsync().ConfigureAwait(false);
-          }
-        }
-      }
-      // </recognitionWithCompressedInputPushStreamAudio>
-    }
-
-    // Speech recognition with audio stream
-    public static async Task RecognitionWithPullAudioStreamAsync() {
-      // Creates an instance of a speech config with specified subscription key and service region.
-      // Replace with your own subscription key and service region (e.g., "westus").
-      var config = SpeechConfig.FromSubscription("YourSubscriptionKey", "YourServiceRegion");
-
-      var stopRecognition = new TaskCompletionSource<int>();
-
-      // Create an audio stream from a wav file.
-      // Replace with your own audio file name.
-      using (var audioInput = Utility.OpenWavFile("test2.wav")) {
-        // Creates a speech recognizer using audio stream input.
-        using (var recognizer = new SpeechRecognizer(config, audioInput)) {
-          // Subscribes to events.
-          recognizer.Recognizing += (s, e) =>
-          {
-            Console.WriteLine($"RECOGNIZING: Text={e.Result.Text}");
-          };
-
-          recognizer.Recognized += (s, e) =>
-          {
-            if (e.Result.Reason == ResultReason.RecognizedSpeech) {
-              Console.WriteLine($"RECOGNIZED: Text={e.Result.Text}");
-            } else if (e.Result.Reason == ResultReason.NoMatch) {
-              Console.WriteLine($"NOMATCH: Speech could not be recognized.");
-            }
-          };
-
-          recognizer.Canceled += (s, e) =>
-          {
-            Console.WriteLine($"CANCELED: Reason={e.Reason}");
-
-            if (e.Reason == CancellationReason.Error) {
-              Console.WriteLine($"CANCELED: ErrorCode={e.ErrorCode}");
-              Console.WriteLine($"CANCELED: ErrorDetails={e.ErrorDetails}");
-              Console.WriteLine($"CANCELED: Did you update the subscription info?");
-            }
-
-            stopRecognition.TrySetResult(0);
-          };
-
-          recognizer.SessionStarted += (s, e) =>
-          {
-            Console.WriteLine("\nSession started event.");
-          };
-
-          recognizer.SessionStopped += (s, e) =>
-          {
-            Console.WriteLine("\nSession stopped event.");
-            Console.WriteLine("\nStop recognition.");
-            stopRecognition.TrySetResult(0);
-          };
-
-          // Starts continuous recognition. Uses StopContinuousRecognitionAsync() to stop recognition.
-          await recognizer.StartContinuousRecognitionAsync().ConfigureAwait(false);
-
-          // Waits for completion.
-          // Use Task.WaitAny to keep the task rooted.
-          Task.WaitAny(new[] { stopRecognition.Task });
-
-          // Stops recognition.
-          await recognizer.StopContinuousRecognitionAsync().ConfigureAwait(false);
-        }
-      }
-    }
-
-    public static async Task RecognitionWithPushAudioStreamAsync() {
-      // Creates an instance of a speech config with specified subscription key and service region.
-      // Replace with your own subscription key and service region (e.g., "westus").
-      var config = SpeechConfig.FromSubscription("YourSubscriptionKey", "YourServiceRegion");
-
-      var stopRecognition = new TaskCompletionSource<int>();
-
-      // Create a push stream
-      using (var pushStream = AudioInputStream.CreatePushStream()) {
-        using (var audioInput = AudioConfig.FromStreamInput(pushStream)) {
-          // Creates a speech recognizer using audio stream input.
-          using (var recognizer = new SpeechRecognizer(config, audioInput)) {
-            // Subscribes to events.
-            recognizer.Recognizing += (s, e) =>
-            {
-              Console.WriteLine($"RECOGNIZING: Text={e.Result.Text}");
-            };
-
-            recognizer.Recognized += (s, e) =>
-            {
-              if (e.Result.Reason == ResultReason.RecognizedSpeech) {
-                Console.WriteLine($"RECOGNIZED: Text={e.Result.Text}");
-              } else if (e.Result.Reason == ResultReason.NoMatch) {
-                Console.WriteLine($"NOMATCH: Speech could not be recognized.");
-              }
-            };
-
-            recognizer.Canceled += (s, e) =>
-            {
-              Console.WriteLine($"CANCELED: Reason={e.Reason}");
-
-              if (e.Reason == CancellationReason.Error) {
-                Console.WriteLine($"CANCELED: ErrorCode={e.ErrorCode}");
-                Console.WriteLine($"CANCELED: ErrorDetails={e.ErrorDetails}");
-                Console.WriteLine($"CANCELED: Did you update the subscription info?");
-              }
-
-              stopRecognition.TrySetResult(0);
-            };
-
-            recognizer.SessionStarted += (s, e) =>
-            {
-              Console.WriteLine("\nSession started event.");
-            };
-
-            recognizer.SessionStopped += (s, e) =>
-            {
-              Console.WriteLine("\nSession stopped event.");
-              Console.WriteLine("\nStop recognition.");
-              stopRecognition.TrySetResult(0);
-            };
-
-            // Starts continuous recognition. Uses StopContinuousRecognitionAsync() to stop recognition.
-            await recognizer.StartContinuousRecognitionAsync().ConfigureAwait(false);
-
-            // open and read the wave file and push the buffers into the recognizer
-            using (BinaryAudioStreamReader reader = Utility.CreateWavReader(Path.GetFullPath("test2.wav"))) {
-              byte[] buffer = new byte[1000];
-              while (true) {
-                var readSamples = reader.Read(buffer, (uint)buffer.Length);
-                if (readSamples == 0) {
-                  break;
-                }
-                pushStream.Write(buffer, readSamples);
-              }
-            }
-            pushStream.Close();
-
-            // Waits for completion.
-            // Use Task.WaitAny to keep the task rooted.
-            Task.WaitAny(new[] { stopRecognition.Task });
-
-            // Stops recognition.
-            await recognizer.StopContinuousRecognitionAsync().ConfigureAwait(false);
-          }
-        }
-      }
-    }
-
-    // Continuous speech recognition with keyword spotting.
-    public static async Task ContinuousRecognitionWithKeywordSpottingAsync() {
-      // Creates an instance of a speech config with specified subscription key and service region.
-      // Replace with your own subscription key and service region (e.g., "westus").
-      var config = SpeechConfig.FromSubscription("YourSubscriptionKey", "YourServiceRegion");
-
-      // Creates an instance of a keyword recognition model. Update this to
-      // point to the location of your keyword recognition model.
-      var model = KeywordRecognitionModel.FromFile("YourKeywordRecognitionModelFile.table");
-
-      // The phrase your keyword recognition model triggers on.
-      var keyword = "YourKeyword";
-
-      var stopRecognition = new TaskCompletionSource<int>();
-
-      // Creates a speech recognizer using microphone as audio input.
-      using (var recognizer = new SpeechRecognizer(config)) {
-        // Subscribes to events.
-        recognizer.Recognizing += (s, e) =>
-        {
-          if (e.Result.Reason == ResultReason.RecognizingKeyword) {
-            Console.WriteLine($"RECOGNIZING KEYWORD: Text={e.Result.Text}");
-          } else if (e.Result.Reason == ResultReason.RecognizingSpeech) {
-            Console.WriteLine($"RECOGNIZING: Text={e.Result.Text}");
-          }
-        };
-
-        recognizer.Recognized += (s, e) =>
-        {
-          if (e.Result.Reason == ResultReason.RecognizedKeyword) {
-            Console.WriteLine($"RECOGNIZED KEYWORD: Text={e.Result.Text}");
-          } else if (e.Result.Reason == ResultReason.RecognizedSpeech) {
-            Console.WriteLine($"RECOGNIZED: Text={e.Result.Text}");
-          } else if (e.Result.Reason == ResultReason.NoMatch) {
-            Console.WriteLine("NOMATCH: Speech could not be recognized.");
-          }
-        };
-
-        recognizer.Canceled += (s, e) =>
-        {
-          Console.WriteLine($"CANCELED: Reason={e.Reason}");
-
-          if (e.Reason == CancellationReason.Error) {
-            Console.WriteLine($"CANCELED: ErrorCode={e.ErrorCode}");
-            Console.WriteLine($"CANCELED: ErrorDetails={e.ErrorDetails}");
-            Console.WriteLine($"CANCELED: Did you update the subscription info?");
-          }
-          stopRecognition.TrySetResult(0);
-        };
-
-        recognizer.SessionStarted += (s, e) =>
-        {
-          Console.WriteLine("\n    Session started event.");
-        };
-
-        recognizer.SessionStopped += (s, e) =>
-        {
-          Console.WriteLine("\n    Session stopped event.");
-          Console.WriteLine("\nStop recognition.");
-          stopRecognition.TrySetResult(0);
-        };
-
-        // Starts recognizing.
-        Console.WriteLine($"Say something starting with the keyword '{keyword}' followed by whatever you want...");
-
-        // Starts continuous recognition using the keyword model. Use
-        // StopKeywordRecognitionAsync() to stop recognition.
-        await recognizer.StartKeywordRecognitionAsync(model).ConfigureAwait(false);
-
-        // Waits for a single successful keyword-triggered speech recognition (or error).
-        // Use Task.WaitAny to keep the task rooted.
-        Task.WaitAny(new[] { stopRecognition.Task });
-
-        // Stops recognition.
-        await recognizer.StopKeywordRecognitionAsync().ConfigureAwait(false);
-      }
-    }
-
-    // Continuous speech recognition assisted with a phrase list.
-    public static async Task ContinuousRecognitionWithFileAndPhraseListsAsync() {
-      // Creates an instance of a speech config with specified subscription key and service region.
-      // Replace with your own subscription key and service region (e.g., "westus").
-      var config = SpeechConfig.FromSubscription("YourSubscriptionKey", "YourServiceRegion");
-
-      var stopRecognition = new TaskCompletionSource<int>();
-
-      // Creates a speech recognizer using file as audio input.
-      // Replace with your own audio file name.
-      using (var audioInput = AudioConfig.FromWavFileInput(@"wreck-a-nice-beach.wav")) {
-        using (var recognizer = new SpeechRecognizer(config, audioInput)) {
-          // Subscribes to events.
-          recognizer.Recognizing += (s, e) =>
-          {
-            Console.WriteLine($"RECOGNIZING: Text={e.Result.Text}");
-          };
-
-          recognizer.Recognized += (s, e) =>
-          {
-            if (e.Result.Reason == ResultReason.RecognizedSpeech) {
-              Console.WriteLine($"RECOGNIZED: Text={e.Result.Text}");
-            } else if (e.Result.Reason == ResultReason.NoMatch) {
-              Console.WriteLine($"NOMATCH: Speech could not be recognized.");
-            }
-          };
-
-          recognizer.Canceled += (s, e) =>
-          {
-            Console.WriteLine($"CANCELED: Reason={e.Reason}");
-
-            if (e.Reason == CancellationReason.Error) {
-              Console.WriteLine($"CANCELED: ErrorCode={e.ErrorCode}");
-              Console.WriteLine($"CANCELED: ErrorDetails={e.ErrorDetails}");
-              Console.WriteLine($"CANCELED: Did you update the subscription info?");
-            }
-
-            stopRecognition.TrySetResult(0);
-          };
-
-          recognizer.SessionStarted += (s, e) =>
-          {
-            Console.WriteLine("\n    Session started event.");
-          };
-
-          recognizer.SessionStopped += (s, e) =>
-          {
-            Console.WriteLine("\n    Session stopped event.");
-            Console.WriteLine("\nStop recognition.");
-            stopRecognition.TrySetResult(0);
-          };
-
-          // Before starting recognition, add a phrase list to help recognition.
-          PhraseListGrammar phraseListGrammar = PhraseListGrammar.FromRecognizer(recognizer);
-          phraseListGrammar.AddPhrase("Wreck a nice beach");
-
-          // Starts continuous recognition. Uses StopContinuousRecognitionAsync() to stop recognition.
-          await recognizer.StartContinuousRecognitionAsync().ConfigureAwait(false);
-
-          // Waits for completion.
-          // Use Task.WaitAny to keep the task rooted.
-          Task.WaitAny(new[] { stopRecognition.Task });
-
-          // Stops recognition.
-          await recognizer.StopContinuousRecognitionAsync().ConfigureAwait(false);
-        }
-      }
-    }
-
-    // Speech recognition with auto detection for source language
-    public static async Task RecognitionWithAutoDetectSourceLanguageAsync() {
-      // Creates an instance of a speech config with specified subscription key and service region.
-      // Replace with your own subscription key and service region (e.g., "westus").
-      var config = SpeechConfig.FromSubscription("YourSubscriptionKey", "YourServiceRegion");
-
-      // Creates an instance of AutoDetectSourceLanguageConfig with the 2 source language candidates
-      // Currently this feature only supports 2 different language candidates
-      // Replace the languages to be the language candidates for your speech. Please see https://docs.microsoft.com/azure/cognitive-services/speech-service/language-support for all supported langauges
-      var autoDetectSourceLanguageConfig = AutoDetectSourceLanguageConfig.FromLanguages(new string[] { "de-DE", "fr-FR" });
-
-      var stopRecognition = new TaskCompletionSource<int>();
-
-      // Creates a speech recognizer using the auto detect source language config, and the file as audio input.
-      // Replace with your own audio file name.
-      using (var audioInput = AudioConfig.FromWavFileInput(Path.GetFullPath("test2.wav"))) {
-        using (var recognizer = new SpeechRecognizer(config, autoDetectSourceLanguageConfig, audioInput)) {
-          // Subscribes to events.
-          recognizer.Recognizing += (s, e) =>
-          {
-            if (e.Result.Reason == ResultReason.RecognizingSpeech) {
-              Console.WriteLine($"RECOGNIZING: Text={e.Result.Text}");
-              // Retrieve the detected language
-              var autoDetectSourceLanguageResult = AutoDetectSourceLanguageResult.FromResult(e.Result);
-              Console.WriteLine($"DETECTED: Language={autoDetectSourceLanguageResult.Language}");
-            }
-          };
-
-          recognizer.Recognized += (s, e) =>
-          {
-            if (e.Result.Reason == ResultReason.RecognizedSpeech) {
-              Console.WriteLine($"RECOGNIZED: Text={e.Result.Text}");
-              // Retrieve the detected language
-              var autoDetectSourceLanguageResult = AutoDetectSourceLanguageResult.FromResult(e.Result);
-              Console.WriteLine($"DETECTED: Language={autoDetectSourceLanguageResult.Language}");
-            } else if (e.Result.Reason == ResultReason.NoMatch) {
-              Console.WriteLine($"NOMATCH: Speech could not be recognized.");
-            }
-          };
-
-          recognizer.Canceled += (s, e) =>
-          {
-            Console.WriteLine($"CANCELED: Reason={e.Reason}");
-
-            if (e.Reason == CancellationReason.Error) {
-              Console.WriteLine($"CANCELED: ErrorCode={e.ErrorCode}");
-              Console.WriteLine($"CANCELED: ErrorDetails={e.ErrorDetails}");
-              Console.WriteLine($"CANCELED: Did you update the subscription info?");
-            }
-
-            stopRecognition.TrySetResult(0);
-          };
-
-          recognizer.SessionStarted += (s, e) =>
-          {
-            Console.WriteLine("\n    Session started event.");
-          };
-
-          recognizer.SessionStopped += (s, e) =>
-          {
-            Console.WriteLine("\n    Session stopped event.");
-            Console.WriteLine("\nStop recognition.");
-            stopRecognition.TrySetResult(0);
-          };
-
-          // Starts continuous recognition. Uses StopContinuousRecognitionAsync() to stop recognition.
-          await recognizer.StartContinuousRecognitionAsync().ConfigureAwait(false);
-
-          // Waits for completion.
-          // Use Task.WaitAny to keep the task rooted.
-          Task.WaitAny(new[] { stopRecognition.Task });
-
-          // Stops recognition.
-          await recognizer.StopContinuousRecognitionAsync().ConfigureAwait(false);
-        }
-      }
-    }
-
-    // Speech recognition with auto detection for source language and custom model
-    public static async Task RecognitionWithAutoDetectSourceLanguageAndCustomModelAsync() {
-      // Creates an instance of a speech config with specified subscription key and service region.
-      // Replace with your own subscription key and service region (e.g., "westus").
-      var config = SpeechConfig.FromSubscription("YourSubscriptionKey", "YourServiceRegion");
-
-      var sourceLanguageConfigs = new SourceLanguageConfig[]
-      {
-                // The endpoint id is optional, if not specified,  the service will use the default model for en-US
-                // Replace the language with your source language candidate. Please see https://docs.microsoft.com/azure/cognitive-services/speech-service/language-support for all supported langauges
-                SourceLanguageConfig.FromLanguage("en-US"),
-
-                // Replace the id with the CRIS endpoint id of your customized model. If the speech is in fr-FR, the service will use the corresponding customized model for speech recognition
-                SourceLanguageConfig.FromLanguage("fr-FR", "YourEndpointId"),
-      };
-
-      // Creates an instance of AutoDetectSourceLanguageConfig with the 2 source language configurations
-      // Currently this feature only supports 2 different language candidates
-      var autoDetectSourceLanguageConfig = AutoDetectSourceLanguageConfig.FromSourceLanguageConfigs(sourceLanguageConfigs);
-
-      var stopRecognition = new TaskCompletionSource<int>();
-
-      // Creates a speech recognizer using the auto detect source language config, and the file as audio input.
-      // Replace with your own audio file name.
-      using (var audioInput = AudioConfig.FromWavFileInput(Path.GetFullPath("test2.wav"))) {
-        using (var recognizer = new SpeechRecognizer(config, autoDetectSourceLanguageConfig, audioInput)) {
-          recognizer.Recognizing += (s, e) =>
-          {
-            if (e.Result.Reason == ResultReason.RecognizingSpeech) {
-              Console.WriteLine($"RECOGNIZING: Text={e.Result.Text}");
-              // Retrieve the detected language
-              var autoDetectSourceLanguageResult = AutoDetectSourceLanguageResult.FromResult(e.Result);
-              Console.WriteLine($"DETECTED: Language={autoDetectSourceLanguageResult.Language}");
-            }
-          };
-
-          recognizer.Recognized += (s, e) =>
-          {
-            if (e.Result.Reason == ResultReason.RecognizedSpeech) {
-              Console.WriteLine($"RECOGNIZED: Text={e.Result.Text}");
-              // Retrieve the detected language
-              var autoDetectSourceLanguageResult = AutoDetectSourceLanguageResult.FromResult(e.Result);
-              Console.WriteLine($"DETECTED: Language={autoDetectSourceLanguageResult.Language}");
-            } else if (e.Result.Reason == ResultReason.NoMatch) {
-              Console.WriteLine($"NOMATCH: Speech could not be recognized.");
-            }
-          };
-
-          recognizer.Canceled += (s, e) =>
-          {
-            Console.WriteLine($"CANCELED: Reason={e.Reason}");
-
-            if (e.Reason == CancellationReason.Error) {
-              Console.WriteLine($"CANCELED: ErrorCode={e.ErrorCode}");
-              Console.WriteLine($"CANCELED: ErrorDetails={e.ErrorDetails}");
-              Console.WriteLine($"CANCELED: Did you update the subscription info?");
-            }
-
-            stopRecognition.TrySetResult(0);
-          };
-
-          recognizer.SessionStarted += (s, e) =>
-          {
-            Console.WriteLine("\n    Session started event.");
-          };
-
-          recognizer.SessionStopped += (s, e) =>
-          {
-            Console.WriteLine("\n    Session stopped event.");
-            Console.WriteLine("\nStop recognition.");
-            stopRecognition.TrySetResult(0);
-          };
-
-          // Starts continuous recognition. Uses StopContinuousRecognitionAsync() to stop recognition.
-          await recognizer.StartContinuousRecognitionAsync().ConfigureAwait(false);
-
-          // Waits for completion.
-          // Use Task.WaitAny to keep the task rooted.
-          Task.WaitAny(new[] { stopRecognition.Task });
-
-          // Stops recognition.
-          await recognizer.StopContinuousRecognitionAsync().ConfigureAwait(false);
-        }
-      }
-    }
-
-    public static async Task KeywordRecognizer() {
-      Console.WriteLine("say something ...");
-      using (var audioInput = AudioConfig.FromDefaultMicrophoneInput()) {
-        using (var recognizer = new KeywordRecognizer(audioInput)) {
-          var model = KeywordRecognitionModel.FromFile("YourKeywordModelFilename.");
-          var result = await recognizer.RecognizeOnceAsync(model).ConfigureAwait(false);
-          Console.WriteLine($"got result reason as {result.Reason}");
-          if (result.Reason == ResultReason.RecognizedKeyword) {
-            var stream = AudioDataStream.FromResult(result);
-
-            await Task.Delay(2000);
-
-            stream.DetachInput();
-            await stream.SaveToWaveFileAsync("AudioFromRecognizedKeyword.wav");
-          } else {
-            Console.WriteLine($"got result reason as {result.Reason}. You can't get audio when no keyword is recognized.");
-          }
-        }
-      }
-    }
-  }
-}
